Test experiments

+++ bin/experiment_test.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import os
+import os, sys
 import subprocess, threading
 import logging as loggers
 import time
@@ -17,16 +17,29 @@ class ParallelExecutor(object):
         self.stderr_output = ""
 
     def run(self, timeout):
-        logging.info(self.cmd)
-        self.process = subprocess.Popen(self.cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        self.stdout_output, self.stderr_output = self.process.communicate(timeout=timeout)
-        logging.info("Process finished")
+        def target():
+            logging.info(self.cmd)
+            self.process = subprocess.Popen(self.cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            self.stdout_output, self.stderr_output = self.process.communicate()
+            logging.info("Thread finished")
 
-        self.process.terminate()
-        time.sleep(1)
+        thread = threading.Thread(target=target)
+        thread.daemon = True
+        thread.start()
+
+        thread.join(timeout)
+        if thread.is_alive():
+            logging.info("Time out, terminating process")
+            self.process.terminate()
+            time.sleep(1)
+            try:
+                self.process.kill()
+            except OSError:
+                pass
+            # thread.join()
         try:
-            self.process.kill()
-        except OSError:
+            thread._stop()
+        except:
             pass
         return self.process.returncode
 
@@ -49,9 +62,10 @@ def test_experiment(path, timeout=60):
 
 if __name__ == '__main__':
     # MNIST
-    test_experiment("mnist/mlp_dropout.py", timeout=30)
+    test_experiment("mnist/mlp_dropout.py", timeout=10)
     test_experiment("mnist/deep_convolution.py", timeout=30)
     # LMs
     test_experiment("lm/baseline_rnnlm.py")
     # Highway networks
     test_experiment("highway_networks/mnist_highway.py")
+    sys.exit(0)
